Reworked footstep sounds to be more generic
