+++ Assets/Scripts/Character/TimedPlayerFootstepSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74f0e7ef83b518f4e8c9c41ca1bb9cda
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
