+++ Assets/Sound/SFX/Scrape.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 632ff49c3e29ee34ca443bdab27bc3d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
